+++ src/Client/ClientSetup.h
@@ -84,6 +84,9 @@ public:
         bypass->modules.push_back(new Module("Character Limit", "char-limit", "Removes the <cl>character limit</c> on input fields"));
         bypass->modules.push_back(new Module("Character Filter", "char-filter", "Allows you to input <cg>any</c> character in input fields"));
 
+        bypass->modules.push_back(new Module("Confirm Practice", "conf-prac", "confirm practice mode help"));
+        bypass->modules.push_back(new Module("Confirm Restart", "conf-res", "confirm restart help"));
+
         #ifdef GEODE_IS_ANDROID
         //bypass->modules.push_back(new Module("Slider Limit Bypass", "slider-limit", "Allows sliders to go beyond the limit of the slider. <cr>Doesn't work for scaling in the editor currently</c>"));
         #endif