+++ src/Hacks/NoWavePulse.cpp
@@ -1,0 +1,18 @@
+/*#include <Geode/Geode.hpp>
+#include <Geode/modify/HardStreak.hpp>
+#include "../Client/Client.h"
+
+using namespace geode::prelude;
+
+class $modify (HardStreak)
+{   
+    void updateStroke(float p0)
+    {
+        // 2.1 bindings don't work sad
+        m_pulseSize = 25;
+
+        log::info("unk: {}", p0);
+
+        HardStreak::updateStroke(p0);
+    }
+};*/