+++ src/Replay/Macro.h
@@ -1,5 +1,5 @@
 #include <Geode/Geode.hpp>
-#include <Geode/modify/EndLevelLayer.hpp>
+#include <Geode/modify/PauseLayer.hpp>
 #include "../Client/Client.h"
 
 using namespace geode::prelude;