+++ src/Replay/Replay.cpp
@@ -1,1 +1,14 @@
-#include "Macro.h"
+/*#include "Macro.h"
+
+class $modify (PauseLayer)
+{
+    virtual TodoReturn customSetup()
+    {
+        PauseLayer::customSetup();
+
+        //auto m = CCMenu::create(CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("GJ_achImage_001.png"), this, nullptr));
+
+        //this->addChild(m);
+    }
+    //TodoReturn setupReplay(gd::string)
+};*/