+++ src/Hacks/ShowLayout.cpp
@@ -1,0 +1,17 @@
+/*#include <Geode/Geode.hpp>
+#include <Geode/modify/PlayLayer.hpp>
+#include "../Client/Client.h"
+
+using namespace geode::prelude;
+
+class $modify (PlayLayer)
+{
+    TodoReturn addObject(GameObject* p0)
+    {
+        //log::info("type: {}", p0->m_objectType == GameObjectType::Decoration);
+
+        //if (p0->m_objectType != GameObjectType::Decoration)
+        //if (false)
+            //PlayLayer::addObject(p0);
+    }
+};*/