+++ src/Hacks/TransitionTimeCustomiser.cpp
@@ -1,9 +1,70 @@
 #include <Geode/Geode.hpp>
 #include <Geode/cocos/layers_scenes_transitions_nodes/CCTransition.h>
 #include <Geode/modify/CCTransitionFade.hpp>
+#include <Geode/modify/CCDirector.hpp>
 #include "../Client/Client.h"
 
 using namespace geode::prelude;
+
+CCScene* getSceneForSel(int i, float f, CCScene* s)
+{
+    AppDelegate::get()->willSwitchToScene(s);
+
+    if (i == 0)
+        return CCTransitionFade::create(f, s);
+    else if (i == 1)
+        return CCTransitionCrossFade::create(f, s);
+    else if (i == 2)
+        return CCTransitionFadeBL::create(f, s);
+    else if (i == 3)
+        return CCTransitionFadeTR::create(f, s);
+    else if (i == 4)
+        return CCTransitionFadeUp::create(f, s);
+    else if (i == 5)
+        return CCTransitionFadeDown::create(f, s);
+    else if (i == 6)
+        return CCTransitionFlipAngular::create(f, s);
+    else if (i == 7)
+        return CCTransitionFlipX::create(f, s);
+    else if (i == 8)
+        return CCTransitionFlipY::create(f, s);
+    else if (i == 9)
+        return CCTransitionZoomFlipAngular::create(f, s);
+    else if (i == 10)
+        return CCTransitionZoomFlipX::create(f, s);
+    else if (i == 11)
+        return CCTransitionZoomFlipY::create(f, s);
+    else if (i == 12)
+        return CCTransitionJumpZoom::create(f, s);
+    else if (i == 13)
+        return CCTransitionMoveInT::create(f, s);
+    else if (i == 14)
+        return CCTransitionMoveInB::create(f, s);
+    else if (i == 15)
+        return CCTransitionMoveInL::create(f, s);
+    else if (i == 16)
+        return CCTransitionMoveInR::create(f, s);
+    else if (i == 17)
+        return CCTransitionRotoZoom::create(f, s);
+    else if (i == 18)
+        return CCTransitionShrinkGrow::create(f, s);
+    else if (i == 19)
+        return CCTransitionSlideInT::create(f, s);
+    else if (i == 20)
+        return CCTransitionSlideInB::create(f, s);
+    else if (i == 21)
+        return CCTransitionSlideInL::create(f, s);
+    else if (i == 22)
+        return CCTransitionSlideInR::create(f, s);
+    else if (i == 23)
+        return CCTransitionSplitRows::create(f, s);
+    else if (i == 24)
+        return CCTransitionSplitCols::create(f, s);
+    else if (i == 25)
+        return CCTransitionTurnOffTiles::create(f, s);
+
+    return nullptr;
+}
 
 class $modify (cocos2d::CCTransitionFade)
 {
@@ -16,6 +77,6 @@ class $modify (cocos2d::CCTransitionFade)
             mod = 0;
         }
 
-        return CCTransitionFade::create(duration * mod, scene);
+        return as<CCTransitionFade*>(getSceneForSel(0, duration * mod, scene));// /*base_cast<CCTransitionFade*>(CCTransitionFlipY::create(duration * mod, scene));  */CCTransitionFade::create(duration * mod, scene);
     }
 };